avoid undefined signed integer overflow
